Add SearchParamsContext for search parameters state handling
SearchParamsContext file has been added to manage the state of search parameters. This context was integrated into the TransientInput component, replacing the previous CoordinatesContext. 'transient' has been changed to 'transientName' in accordance with the new context. This context is more generic, paving a way for reuse in different components. This modification enhances the overall code modularity.