Delete `loadGCNRecord.js`, rename `loadAtelRecord.js` to `loadRecords.js`
In order to consolidate and improve the process of data loading and date parsing for 'atel' and 'gcn' data sets, we have deleted the redundant `loadGCNRecord.js` file and incorporated its functionalities into a renamed `loadRecords.js` file. This streamlining is part of our continued efforts to maintain clean, lean code.