Update README references and app title
Additional relevant references have been added to README.md. In addition, the title of the application has been changed in the HTML metadata. These edits aim to provide users and developers with more complete and accurate project information, and ensure that the application's branding is consistent across all its functionalities.